+++ install.sh
@@ -29,11 +29,6 @@ echo ""
 deactivate
 
 echo ""
-echo "Installing python dependencies for mfg_tool"
-echo ""
-python3 -m pip install -r ${ESP_MATTER_PATH}/tools/mfg_tool/requirements.txt
-
-echo ""
 echo "Installing python dependencies for Matter"
 echo ""
 python3 -m pip install -r ${ESP_MATTER_PATH}/requirements.txt